add traj mode to train pi0 based on dobot dataset

+++ src/openpi/policies/policy_config.py
@@ -53,7 +53,8 @@ def create_trained_policy(
     checkpoint_dir = download.maybe_download(str(checkpoint_dir))
 
     logging.info("Loading model...")
-    model = train_config.model.load(_model.restore_params(checkpoint_dir / "params", dtype=jnp.bfloat16))
+    # model = train_config.model.load(_model.restore_params(checkpoint_dir / "params"))
+    model = train_config.model.load(_model.restore_params(checkpoint_dir / "params"))
 
     data_config = train_config.data.create(train_config.assets_dirs, train_config.model)
     if norm_stats is None:
